Refactor: move FleetGateway interface on the user package's side
